backend/requirements.txt
fastapi[standard]>=0.110.0
google-genai>=0.1.2
python-dotenv>=1.0.1
uvicorn>=0.29.0
